菜单模块添加按钮功能

+++ product-server/src/main/java/com/product/sysrole/bean/SysRole.java
@@ -1,15 +1,18 @@
 package com.product.sysrole.bean;
 
 public class SysRole {
-    private Integer id;
+    private Long id;
 
+    /**
+     * 角色名称
+     */
     private String name;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
